+++ grammar/interfaces.go
@@ -10,9 +10,8 @@ type Valuer interface {
 type Canonizer interface {
 	// canonical function would return something only for nodes that do
 	// contribute to canonical representation. For other nodes the return
-	// value is an empty canonical structure. The second value indicates if
-	// the canonical is a hybrid or not.
-	canonical() (*Canonical, bool)
+	// value is an empty canonical structure.
+	canonical() *Canonical
 }
 
 type Poser interface {